# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: PitMaster AI - BBQ Planning & Monitoring API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /grills:
    get:
      operationId: listGrills
      tags: [grills]
      summary: List all grill profiles
      responses:
        "200":
          description: List of grill profiles
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Grill"
    post:
      operationId: createGrill
      tags: [grills]
      summary: Create a new grill profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateGrillBody"
      responses:
        "201":
          description: Created grill
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Grill"
  /grills/{id}:
    get:
      operationId: getGrill
      tags: [grills]
      summary: Get a grill profile by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Grill profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Grill"
        "404":
          description: Not found
    patch:
      operationId: updateGrill
      tags: [grills]
      summary: Update a grill profile
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateGrillBody"
      responses:
        "200":
          description: Updated grill
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Grill"
        "404":
          description: Not found
    delete:
      operationId: deleteGrill
      tags: [grills]
      summary: Delete a grill profile
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /grills/{id}/stats:
    get:
      operationId: getGrillStats
      tags: [grills]
      summary: Get cook stats for a specific grill
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Grill stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GrillStats"
...
  /recipes:
    get:
      operationId: listRecipes
      tags: [recipes]
      summary: List all recipes
      parameters:
        - name: category
          in: query
          schema:
            type: ["string", "null"]
        - name: search
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: List of recipes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Recipe"
    post:
      operationId: createRecipe
      tags: [recipes]
      summary: Create a new recipe
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRecipeBody"
      responses:
        "201":
          description: Created recipe
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Recipe"
  /recipes/{id}:
    get:
      operationId: getRecipe
      tags: [recipes]
      summary: Get a recipe by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Recipe
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Recipe"
        "404":
          description: Not found
    patch:
      operationId: updateRecipe
      tags: [recipes]
      summary: Update a recipe
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateRecipeBody"
      responses:
        "200":
          description: Updated recipe
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Recipe"
    delete:
      operationId: deleteRecipe
      tags: [recipes]
      summary: Delete a recipe
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /recipes/{id}/favorite:
    patch:
      operationId: toggleRecipeFavorite
      tags: [recipes]
      summary: Toggle recipe favorite status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated recipe
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Recipe"
  /temperature/upload:
    post:
      operationId: uploadTemperatureData
      tags: [temperature]
      summary: Upload temperature data from thermometer apps
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TemperatureUploadBody"
      responses:
        "201":
          description: Uploaded temperature readings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TemperatureUploadResult"
  /temperature/readings:
    get:
      operationId: listTemperatureReadings
      tags: [temperature]
      summary: Get temperature readings for a cook session
      parameters:
        - name: cookId
          in: query
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Temperature readings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TemperatureReading"
...
components:
  schemas:
...
    Grill:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        type:
          type: string
          description: "charcoal, gas, pellet, electric, smoker, kamado, etc."
        brand:
          type: ["string", "null"]
        model:
          type: ["string", "null"]
        cookingSurfaceSqIn:
          type: ["number", "null"]
        maxTempF:
          type: ["number", "null"]
        notes:
          type: ["string", "null"]
        imageUrl:
          type: ["string", "null"]
        totalCooks:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - type
        - brand
        - model
        - cookingSurfaceSqIn
        - maxTempF
        - notes
        - imageUrl
        - totalCooks
        - createdAt
        - updatedAt
    CreateGrillBody:
      type: object
      properties:
        name:
          type: string
        type:
          type: string
        brand:
          type: ["string", "null"]
        model:
          type: ["string", "null"]
        cookingSurfaceSqIn:
          type: ["number", "null"]
        maxTempF:
          type: ["number", "null"]
        notes:
          type: ["string", "null"]
        imageUrl:
          type: ["string", "null"]
      required:
        - name
        - type
    UpdateGrillBody:
      type: object
      properties:
        name:
          type: ["string", "null"]
        type:
          type: ["string", "null"]
        brand:
          type: ["string", "null"]
        model:
          type: ["string", "null"]
        cookingSurfaceSqIn:
          type: ["number", "null"]
        maxTempF:
          type: ["number", "null"]
        notes:
          type: ["string", "null"]
        imageUrl:
          type: ["string", "null"]
    GrillStats:
      type: object
      properties:
        grillId:
          type: integer
        totalCooks:
          type: integer
        totalHours:
          type: number
        avgCookDurationMinutes:
          type: number
        mostCookedFood:
          type: ["string", "null"]
        avgTargetTempF:
          type: ["number", "null"]
      required:
        - grillId
        - totalCooks
        - totalHours
        - avgCookDurationMinutes
        - mostCookedFood
        - avgTargetTempF
...
    Recipe:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        category:
          type: string
          description: "beef, pork, chicken, fish, vegetables, etc."
        ingredients:
          type: string
        instructions:
          type: string
        cookTempF:
          type: ["number", "null"]
        targetTempF:
          type: ["number", "null"]
        estimatedTimeMinutes:
          type: ["integer", "null"]
        servings:
          type: ["integer", "null"]
        isFavorite:
          type: boolean
        imageUrl:
          type: ["string", "null"]
        tags:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - description
        - category
        - ingredients
        - instructions
        - cookTempF
        - targetTempF
        - estimatedTimeMinutes
        - servings
        - isFavorite
        - imageUrl
        - tags
        - createdAt
        - updatedAt
    CreateRecipeBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: ["string", "null"]
        category:
          type: string
        ingredients:
          type: string
        instructions:
          type: string
        cookTempF:
          type: ["number", "null"]
        targetTempF:
          type: ["number", "null"]
        estimatedTimeMinutes:
          type: ["integer", "null"]
        servings:
          type: ["integer", "null"]
        imageUrl:
          type: ["string", "null"]
        tags:
          type: ["string", "null"]
      required:
        - title
        - category
        - ingredients
        - instructions
    UpdateRecipeBody:
      type: object
      properties:
        title:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        category:
          type: ["string", "null"]
        ingredients:
          type: ["string", "null"]
        instructions:
          type: ["string", "null"]
        cookTempF:
          type: ["number", "null"]
        targetTempF:
          type: ["number", "null"]
        estimatedTimeMinutes:
          type: ["integer", "null"]
        servings:
          type: ["integer", "null"]
        imageUrl:
          type: ["string", "null"]
        tags:
          type: ["string", "null"]
    TemperatureReading:
      type: object
      properties:
        id:
          type: integer
        cookId:
          type: integer
        probeNumber:
          type: integer
        probeName:
          type: ["string", "null"]
        tempF:
          type: number
        recordedAt:
          type: string
          format: date-time
        source:
          type: string
          description: "manual, meater, thermoworks, inkbird, govee, csv"
      required:
        - id
        - cookId
        - probeNumber
        - probeName
        - tempF
        - recordedAt
        - source
    TemperatureUploadBody:
      type: object
      properties:
        cookId:
          type: integer
        source:
          type: string
          description: "meater, thermoworks, inkbird, govee, csv, manual"
        readings:
          type: array
          items:
            $ref: "#/components/schemas/TemperatureReadingInput"
      required:
        - cookId
        - source
        - readings
    TemperatureReadingInput:
      type: object
      properties:
        probeNumber:
          type: integer
        probeName:
          type: ["string", "null"]
        tempF:
          type: number
        recordedAt:
          type: string
          format: date-time
      required:
        - probeNumber
        - tempF
        - recordedAt
    TemperatureUploadResult:
      type: object
      properties:
        inserted:
          type: integer
        cookId:
          type: integer
      required:
        - inserted
        - cookId
...