# lib/api-zod/src/generated/types/grillStats.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * PitMaster AI - BBQ Planning & Monitoring API
 * OpenAPI spec version: 0.1.0
 */

export interface GrillStats {
  grillId: number;
  totalCooks: number;
  totalHours: number;
  avgCookDurationMinutes: number;
  /** @nullable */
  mostCookedFood: string | null;
  /** @nullable */
  avgTargetTempF: number | null;
}