# lib/api-zod/src/generated/types/temperatureReading.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * PitMaster AI - BBQ Planning & Monitoring API
 * OpenAPI spec version: 0.1.0
 */

export interface TemperatureReading {
  id: number;
  cookId: number;
  probeNumber: number;
  /** @nullable */
  probeName: string | null;
  tempF: number;
  recordedAt: Date;
  /** manual, meater, thermoworks, inkbird, govee, csv */
  source: string;
}