# lib/api-zod/src/generated/types/updateGrillBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * PitMaster AI - BBQ Planning & Monitoring API
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateGrillBody {
  /** @nullable */
  name?: string | null;
  /** @nullable */
  type?: string | null;
  /** @nullable */
  brand?: string | null;
  /** @nullable */
  model?: string | null;
  /** @nullable */
  cookingSurfaceSqIn?: number | null;
  /** @nullable */
  maxTempF?: number | null;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  imageUrl?: string | null;
}